made csv reporter roll like log files

+++ metrics-core/src/main/java/com/yammer/metrics/reporting/CsvReporter.java
@@ -1,9 +1,7 @@
 package com.yammer.metrics.reporting;
 
-import com.yammer.metrics.core.*;
-import com.yammer.metrics.util.MetricPredicate;
-
 import java.io.File;
+import java.io.FileFilter;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.PrintStream;
@@ -13,11 +11,30 @@ import java.util.Map.Entry;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
+import com.yammer.metrics.Metrics;
+import com.yammer.metrics.core.CounterMetric;
+import com.yammer.metrics.core.GaugeMetric;
+import com.yammer.metrics.core.HistogramMetric;
+import com.yammer.metrics.core.MeterMetric;
+import com.yammer.metrics.core.Metric;
+import com.yammer.metrics.core.MetricName;
+import com.yammer.metrics.core.MetricsRegistry;
+import com.yammer.metrics.core.TimerMetric;
+import com.yammer.metrics.util.MetricPredicate;
+
 public class CsvReporter extends AbstractPollingReporter {
+
+    private static class StreamPair{
+      String file;
+      PrintStream stream;
+    }
+  
     private final MetricPredicate predicate;
     private final File outputDir;
-    private final Map<MetricName, PrintStream> streamMap;
+    private final Map<MetricName, StreamPair> streamMap;
     private long startTime;
+    private TimeUnit _rollDuration;
+    private File _currentFile;
 
     public CsvReporter(File outputDir,
                        MetricsRegistry metricsRegistry,
@@ -25,25 +42,104 @@ public class CsvReporter extends AbstractPollingReporter {
         super(metricsRegistry, "csv-reporter");
         this.outputDir = outputDir;
         this.predicate = predicate;
-        this.streamMap = new HashMap<MetricName, PrintStream>();
+        this.streamMap = new HashMap<MetricName, StreamPair>();
         this.startTime = 0L;
+        this._rollDuration = TimeUnit.HOURS;
+        _currentFile = null;
+        outputDir.mkdirs();
     }
 
     public CsvReporter(File outputDir, MetricsRegistry metricsRegistry)
             throws Exception {
         this(outputDir, metricsRegistry, MetricPredicate.ALL);
     }
+    
+    public void setRollDuration(TimeUnit duration){
+        _rollDuration = duration;
+    }
+    
+    private static long getFileTime(File file){
+      String name = file.getName();
+      int idx = name.lastIndexOf('-');
+      int idx2 = name.lastIndexOf(".csv");
+      String numberString = name.substring(idx+1,idx2);
+      return Long.parseLong(numberString);
+    }
+    
+    private File getFile(MetricName metricName){
+        final String mname = metricName.getName();
+        
+        File[] files = outputDir.listFiles(new FileFilter(){
+
+          @Override
+          public boolean accept(File pathname) {
+            String filename = pathname.getName();
+            if (filename.startsWith(mname+"-") && filename.endsWith(".csv")){
+              return true;
+            }
+            return false;
+          }
+          
+        });
+        
+        long currentTime = System.nanoTime();
+        if (_currentFile == null){
+          if (files.length==0){
+            String filename = metricName.getName()+"-"+currentTime+".csv";
+            _currentFile = new File(outputDir,filename);
+            return _currentFile;
+          }
+          
+          // get the latest file
+          long maxTime = 0;
+            
+          for (File file : files){
+            long time = getFileTime(file);
+            if (time > maxTime){
+              maxTime = time;
+              _currentFile = file;
+            }
+          }
+        }
+        
+        long fileTime = getFileTime(_currentFile);
+        long duration = currentTime - fileTime;
+          
+        long maxDuration = _rollDuration.toNanos(1);
+          
+        if (duration > maxDuration){
+         String filename = metricName.getName()+"-"+currentTime+".csv";
+         _currentFile = new File(outputDir,filename);
+        }
+        return _currentFile;
+    }
+    
+    
 
     private PrintStream getPrintStream(MetricName metricName, Metric metric)
             throws IOException {
-        PrintStream stream;
+        StreamPair streamPair;
         synchronized (streamMap) {
-            stream = streamMap.get(metricName);
-            if (stream == null) {
-                final File newFile = new File(outputDir, metricName.getName() + ".csv");
-                if (newFile.createNewFile()) {
+            streamPair = streamMap.get(metricName);
+            
+            if (streamPair != null){
+              File f = getFile(metricName);
+              if (!streamPair.file.equals(f.getName())){
+                streamPair.stream.close();
+                streamPair = null;
+              }
+            }
+            if (streamPair == null) {
+                final File newFile = getFile(metricName);
+                
+                streamPair = new StreamPair();
+                streamPair.file = newFile.getName();
+                
+                PrintStream stream;
+                if (!newFile.exists() && newFile.createNewFile()) {
                     stream = new PrintStream(new FileOutputStream(newFile));
-                    streamMap.put(metricName, stream);
+                    streamPair.stream = stream;
+                    streamMap.put(metricName, streamPair);
                     if (metric instanceof GaugeMetric<?>) {
                         stream.println("# time,value");
                     } else if (metric instanceof CounterMetric) {
@@ -60,7 +156,7 @@ public class CsvReporter extends AbstractPollingReporter {
                 }
             }
         }
-        return stream;
+        return streamPair.stream;
     }
 
     @Override
@@ -136,9 +232,38 @@ public class CsvReporter extends AbstractPollingReporter {
         try {
             super.shutdown();
         } finally {
-            for (PrintStream out : streamMap.values()) {
-                out.close();
+            for (StreamPair out : streamMap.values()) {
+                out.stream.close();
             }
         }
+    }
+    
+    public static void main(String[] args) throws Exception{
+      
+     final CounterMetric counterMetric = Metrics.newCounter(new MetricName("group","counter","test"));
+      
+      final int max = 100;
+      Thread t = new Thread(new Runnable(){
+        public void run(){
+          for (int i=0;i<max;++i){
+            counterMetric.inc();
+            
+            try {
+              Thread.sleep(1000);
+            } catch (InterruptedException e) {
+              e.printStackTrace();
+            }
+          }
+        }
+      });
+      
+      t.start();
+      
+      MetricsRegistry metricsRegistry = Metrics.defaultRegistry();
+      CsvReporter reporter = new CsvReporter(new File("/tmp/metrics-test"), metricsRegistry);
+      reporter.setRollDuration(TimeUnit.SECONDS);
+      
+      reporter.start(500, TimeUnit.MILLISECONDS);
+      t.join();
     }
 }
